requirements.txt
# Core Flask Framework
Flask==3.0.0
python-dotenv==1.0.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL adapter (remove if using SQLite only)

# Authentication & Security
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
Flask-Cors==4.0.0

# Validation & Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==1.0.0

# HTTP Requests
requests==2.31.0

# Production Server
gunicorn==21.2.0

Flask-Limiter==3.5.0